refactor(all): 클래스 네이밍, 패키지 정리
