+++ src/main/java/baseball/view/OutputView.java
@@ -2,7 +2,7 @@ package baseball.view;
 
 import java.util.List;
 
-public final class OutputHandler {
+public final class OutputView {
     public void printMessageForData(List<Integer> data) {
         OutputMsgContainer.printMessage(data);
     }
