+++ src/test/java/baseball/utils/RandomNumGeneratorTest.java
@@ -3,6 +3,7 @@ package baseball.utils;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.domain.RandomNumGenerator;
 import org.junit.jupiter.api.Test;
 
 class RandomNumGeneratorTest {
