+++ src/test/java/baseball/view/InputViewTest.java
@@ -6,14 +6,14 @@ import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class InputHandlerTest extends NsTest {
-    private InputHandler inputHandler = new InputHandler();
+public class InputViewTest extends NsTest {
+    private InputView inputView = new InputView();
 
     @Test
     @DisplayName("유저에게 입력받는 세자리 수가 잘 들어오는지 테스트한다")
     void readGameInputTest() {
         super.run("123");
-        assertThat(inputHandler.readGameInput())
+        assertThat(inputView.readGameInput())
                 .isEqualTo(123);
     }
 
@@ -21,7 +21,7 @@ public class InputHandlerTest extends NsTest {
     @DisplayName("유저에게 입력받는 컨트롤 값이 잘 들어오는지 테스트한다")
     void readGameControlInputTest() {
         super.run("1");
-        assertThat(inputHandler.readGameControlInput())
+        assertThat(inputView.readGameControlInput())
                 .isEqualTo(1);
     }
